app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "minsearch>=0.0.7",
    "openai>=2.14.0",
    "pydantic-ai>=1.39.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]
